Skip swapping when in the sequence of 2-byte extended scancodes
Before this commit, swapping would be done for each byte of extended
scancodes, breaking the sequence.
By ignoring 0xE0 and the next byte, prevent this type of destruction.

Actually there are a few extended scancodes which consist of three or
more bytes.
However, now this module doesn't handle those extended scancodes and
there may be unexpected effects depending on your keyboard.

If there are no extended scancodes that include KEY_CAPSLOCK and
KEY_LEFTCTRL in your environment, this module should be safe and you
don't have to worry about that type of destruction.